mwcc: compiler host
